requirements.txt
# Core ML Libraries
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pandas==2.0.3

# Image Processing
pillow==10.0.0
opencv-python==4.8.0.74

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Model Evaluation
scikit-learn==1.3.0

# Web Frameworks
streamlit==1.28.0
gradio==4.0.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1

# For saving/loading
h5py==3.9.0